chore(cleanup): refactor code for cleanups
