--- src/graphql/User.ts
@@ -1,11 +1,0 @@
-import { objectType } from "nexus";
-
-export const User = objectType({
-    name: "User",
-    definition(t) {
-        t.nonNull.int("id");
-        t.nonNull.string("name");
-        t.nonNull.string("email");
-        t.nonNull.string("password");
-    }
-});
