+++ src/graphql/index.ts
@@ -1,3 +1,3 @@
 export * from "./Track";
-export * from "./User";
 export * from "./Auth";
+export * from "./TypeDefinitions";
