+++ src/index.ts
@@ -1,6 +1,7 @@
 import { server } from "./server";
 
-const port = process.env.PORT; // || 3000;
+const port = process.env.PORT || 3000;
+// const port = process.env.PORT; // || 3000;
 
 server.listen({ port }).then(({ url }) => {
     console.log(`🚀  Server ready at ${url}`);
