+++ src/server.ts
@@ -7,6 +7,5 @@ import { context } from "./context";
 export const server = new ApolloServer({
     schema,
     context,
-    // introspection: process.env.NODE_ENV !== "production",
     plugins: [ApolloServerPluginLandingPageGraphQLPlayground()],
 });
